+++ db/models/ssh_key.py
@@ -15,7 +15,7 @@ class SshKeyModel(Base):
     id: Mapped[int] = mapped_column(Integer(), primary_key=True, autoincrement=True)
     title: Mapped[Optional[str]] = mapped_column(String())
     ssh_key: Mapped[str] = mapped_column(String())
-    created_by_id: Mapped[int] = mapped_column(ForeignKey("users.id"))
+    created_by_id: Mapped[int] = mapped_column(ForeignKey("users.id", ondelete="CASCADE"))
     created_by: Mapped["UserModel"] = relationship("UserModel",
                                                    foreign_keys="SshKeyModel.created_by_id")
     created_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.now)